+++ NEWS.md
@@ -5,6 +5,7 @@
 ## major functionality 
 
 * `tmd_to_x3p` allows an import of a file in TMD format (natively supported by GelSight instruments) to x3p
+* `x3p_bin_stripes` color-shade the mask by row or column-baed gradient
 
 ## minor functionality